Create basic_code
Above code is a basic blockchian python code quoted from "http://ecomunsing.com/build-your-own-blockchain"

In this algorithm, additional modifications are required.
First, the algorithm does not have a consensus algorithm part.
Second, the mining node is not set.  (network setup is omitted)
Third, Tx generation occurs automatically. (In fact, it must be synchronized via inter-node broadcasting)
Finally, if the block is not created, the state should not be updated. In this code, However, state is updated immediately when Tx is valid.